+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.3.0] - 2026-03-16
+
+### Added
+
+- Add option to include system drivers
+- Add option to only include system drivers
+
+### Fixed
+
+- Improve import driver match device option
+
 ## [2.2.1] - 2026-02-26
 
 ### Fixed